Rename project
Update project and rename files to fast-api-oidc or fast_api_oidc
