+++ src/fast_api_oidc/auth/multi_provider_auth_middleware.py
@@ -4,10 +4,10 @@ from fastapi.responses import JSONResponse
 import jwt
 from starlette.types import ASGIApp
 from typing import List, Dict, Union, Optional
-from fast_api_jwt_middleware.auth.auth_middleware import AuthMiddleware
-from fast_api_jwt_middleware.cache.cache_protocol_contract import CacheProtocol
-from fast_api_jwt_middleware.context_holder import request_context
-from fast_api_jwt_middleware.logger.logger_protocol_contract import LoggerProtocol
+from fast_api_oidc.auth.auth_middleware import AuthMiddleware
+from fast_api_oidc.cache.cache_protocol_contract import CacheProtocol
+from fast_api_oidc.context_holder import request_context
+from fast_api_oidc.logger.logger_protocol_contract import LoggerProtocol
 
 class MultiProviderAuthMiddleware(AuthMiddleware):
     '''
