+++ src/fast_api_oidc/logger/logger_protocol_contract.py
