+++ tests/test_token_cache_singleton.py
@@ -3,7 +3,7 @@ import sys
 import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
-from fast_api_jwt_middleware.cache.token_cache_singleton import TokenCacheSingleton
+from fast_api_oidc.cache.token_cache_singleton import TokenCacheSingleton
 
 class TestTokenCacheSingleton(unittest.TestCase):
